requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==5.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.2
mdurl==0.1.2
narwhals==2.11.0
numpy==2.2.6
openai==1.52.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==24.2
pandas==2.2.3
pillow==10.4.0
playwright==1.56.0
protobuf==5.29.5
pyarrow==22.0.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==24.2.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.8.0
pytz==2025.2
pyzbar==0.1.9
referencing==0.37.0
requests==2.32.3
requests-file==3.0.1
rich==13.9.4
rpds-py==0.28.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.38.0
streamlit-ace==0.1.1
tenacity==8.5.0
tldextract==5.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
watchdog==4.0.2
